:construction: fetch data
This code needs serious refactoring! It might be
better for me to abstract the entire "data fetching"
part of the code into separate functions that can
go inside of a "utils" folder.

Also, I've decided to store the auth token inside
of a cookie. Yes, I know that's not a good idea and
poses huge security issues, but since this is just
a small personal project, I figured it's
nbd. But my reason for doing this is because
PetFinder's expiration date for the token is
extremely short (like an hour), and I want the
client to automatically be able to get a new
token when it expires.

Also, yes, my token refetching implementation needs
a huge refactor bc it makes zero sense,
I'm not done with it yet lol.