:sparkles: implement auth token refresh
I put the auth token refresh implementation inside of its own module,
and split it up into multiple functions. This could've been made
into a class, so I might do that later. Annnd, there might be a cleaner
way to do this (of course), buuut it gets the job done so I'll keep it
this way for now.

Also, I changed the flags for the cookies because cookies were
not storing on mobile devices (Chrome iOS and Chrome Android). But since
I changed them, everything now works smoothly.