:sparkles: create option component
This is the component that allows users to more narrowly filter the
type of animal they want to see. Species, age, size, gender, good with,
and environment are the possible "option types" as of now. I was also
considering making this into a HOC, because it only wraps around the
React-Select component for now, but I was thinking that it could maybe wrap
around other types of components as well (like a sliding range component).