:sparkles: add fn that parses search options
By "search options," I mean the possible options that the user can
select from the React-Search component. The React-Search component needs
the list of options to be in a certain shape (an object with a "value"
and "label" property), so this function handles that. I also added the
ability for me, the dev, to change the option's name via a prop on the
Option component if I want, and this function handles that as well
(though it's optional).