:sparkles: parse query params and set lastSearchLocation
I included a function to parse the text of the parameters so they
could be in the correct "shape"/way before being put into the actual query
parameter for the request. If the locations are coordinates, I put
them in a form that the API can understand, and same goes for if the
location is a number or string. Same goes for the selected species.
Some of their values have ampersands and commas, which I took out.

I also set the last searched location inside of localStorage, which
will be used as a default location if no location is provided by the
user inside the input elements.