:sparkles: display info of pet inside modal
When the modal opens, display the info of the pet with the ID of the
selected card. The styles are extremely bare for now, I just needed to
display some information to have some actual data inside the modal and
make sure it's "logically" running smoothly.

Clicking either the semi-transparent overlay OR the modal's close
button will close the modal.