:sparkles: create select component
The select component contains an HTML select element that deals with the
selection of the pet species. It parses text to display the option
names in a particular way (whether I just want to capitalize the labels,
pluralize them, etc). It also displays an icon that corresponds to the
currently selected species.