:lipstick: change dog icon to be default icon
My logic behind this was that 'dog' was the first option inside of the
select, so if the user hasn't selected a species, then inevitably
'dog' will be initially displayed inside of the select element (it may
not TECHNICALLY be "selected," but it will still appear as the first
choice). I'll change this later if my logic is proven to be false.