:sparkles::construction: create search page
Oou wee, this component is a mess and I can probably break it down into
smaller components and just... "modularize" it in some way where it's
not this damn massive and messy. Definitely a WIP.

But the Search page is where the search results for the pets appear,
and it also includes search filters that are configurable for the user.

I plan on adding pagination, functionality for the extra search filters,
responsive design, and changing the implementation so that obtaining
search queries from the user is a more easier and streamlined process,
bc IMO it's a bit all over the place.