:recycle: make animal data undefined by default
So that there can be a clear distinction between whether or not an
API request has been made. Even if a request was made and there were
no found results, an array (albeit empty) will STILL be returned. Vs
if there was an error or no request at all, the data will still be
UNDEFINED. I use this logic to determine what to display to the user.
I'll either display an error message, the returned data, or some
text prompting the user to search for something.