:recycle: implement lastSearchedLocation and proper parsing
Now the user's last location they searched for is stored in localStorage
and used as back up if they don't have any location input OR they
haven't provided their location with the geoLocation API.

I've also extracted the location text parsing into its own function,
and added some extra functionality so that it works for state
abbreviations, "city, state", and just states. And if it's a zip code/
not a string, it ofc doesn't parse it.

One could argue that doing all of that is completely pointless, but I did it
purely from an aesthetic POV; I want the displayed location text
to be capitalized... idk why lol.