:recycle::sparkles: store last selected species in local storage
This is needed for the automatic API requests made when the Search page
mounts. If the user refreshes the page, instead of the API returning animals of
ALL species (since the selected species state is initially an empty
string), it instead takes the last searched species and
uses that instead.