:recycle::wheelchair: refactor DistanceInput component for a11y
I wanted the "miles" text to also be detectable by screen readers, but
at first it was not since it was a pseudoelement.

Because I changed it from a pseudoelement to a span tag, I'd
decided to go ahead and make the DistanceInput more customizable to
replicate the style as if the word "miles" just as if it was still a
pseudoelement.